refactor(models): models/mitre/ package - rehome CWEEntry out of models/cwe.py (#161)
Stands up models/mitre/ as the MITRE weakness-taxonomy domain, sibling to
models/nvd/ (NVD/scoring) and distinct from the OAM asset graph. Per-class
files: models/mitre/cwe.py holds CWEEntry, moved from models/cwe.py.

Re-exports keep 'from models import CWEEntry' working unchanged; the two direct
importers (models/__init__ and tools/cwe_data) now point at models.mitre. Also
corrects the models/__init__ doc-table rows that slice 1 (1045c7a) left stale -
Severity/CveEntry under models.nvd, the primitives row, CWEEntry under
models.mitre. Full clean stack verified green before commit (ruff, mypy 0,
pylint 9.92, pytest, bandit, diff-cover 100%).
